# docs/api/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Ollamify External API (static snapshot)",
    "version": "1.0.0",
    "description": "Static snapshot of Ollamify external API. For the live docs (when the stack is running) use: http://localhost/api/docs"
  },
  "servers": [
    {
      "url": "http://localhost/api",
      "description": "Local gateway (default)"
    },
    {
      "url": "/api",
      "description": "Gateway (relative)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Use JWT (web login) or API key in the format: Bearer <token>"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": { "type": "string", "description": "Error message" },
          "code": { "type": "string", "description": "Error code" },
          "details": { "type": "string", "description": "Detailed error information" }
        }
      },
...
      "VoiceInfo": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "example": "omnivoice-he" },
          "gender": { "type": "string", "example": "young" },
          "language": { "type": "string", "example": "ru" },
          "description": { "type": "string" }
        }
      },
      "TTSRequest": {
        "type": "object",
        "required": ["text"],
        "properties": {
          "text": { "type": "string", "maxLength": 1000, "example": "Привет!" },
          "voice": { "type": "string", "enum": ["omnivoice-ru", "omnivoice-en", "omnivoice-he"], "default": "omnivoice-ru" },
          "speed": { "type": "number", "minimum": 0.5, "maximum": 2.0, "default": 0.65, "example": 0.65 },
          "sample_rate": { "type": "integer", "enum": [8000, 24000, 48000], "default": 24000, "example": 24000 },
          "format": { "type": "string", "enum": ["wav"], "default": "wav" },
          "language": { "type": "string", "enum": ["ru", "en", "he"], "default": "ru" }
        }
      },
      "TTSResponse": {
        "type": "object",
        "properties": {
          "audio_base64": { "type": "string", "description": "Base64-encoded WAV" },
          "format": { "type": "string", "example": "wav" },
          "sample_rate": { "type": "integer", "example": 24000 },
          "duration_ms": { "type": "integer", "example": 1200 }
        }
      },
...
    }
  },
  "security": [
    { "ApiKeyAuth": [] }
  ],
  "paths": {
...
    "/tts/voices": {
      "get": {
        "tags": ["TTS"],
        "summary": "List available voices",
        "responses": {
          "200": { "description": "Voices", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/VoiceInfo" } } } } },
          "503": { "description": "Service unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/tts/synthesize": {
      "post": {
        "tags": ["TTS"],
        "summary": "Synthesize speech (base64 WAV)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TTSRequest" } } } },
        "responses": {
          "200": { "description": "TTS response", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TTSResponse" } } } },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Service unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/tts/synthesize/stream": {
      "post": {
        "tags": ["TTS"],
        "summary": "Synthesize speech (binary WAV)",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TTSRequest" } } } },
        "responses": {
          "200": { "description": "WAV audio", "content": { "audio/wav": { "schema": { "type": "string", "format": "binary" } } } },
          "400": { "description": "Bad request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
          "503": { "description": "Service unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
  }
}